# docs/ai-handoff-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WoC AI Handoff",
    "description": "GPT가 WoC의 Claude MAIN/SUB에게 작업을 지시하고 완료 보고를 읽기 위한 최소 API. ChatGPT Custom GPT의 Action으로 그대로 사용 가능.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "https://www.woc.today" }
  ],
  "paths": {
    "/api/ai-handoff/task": {
      "post": {
        "operationId": "create_ai_task",
        "summary": "CLAUDE_MAIN 또는 CLAUDE_SUB에게 새 TASK를 지시한다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["assignee"],
                "properties": {
                  "assignee": {
                    "type": "string",
                    "enum": ["CLAUDE_MAIN", "CLAUDE_SUB"],
                    "description": "이 TASK를 처리할 담당자."
                  },
                  "title": { "type": "string", "description": "TASK 제목 (최대 200자)." },
                  "content": { "type": "string", "description": "TASK 상세 지시 내용 (최대 8000자)." },
                  "scope": {
                    "type": "object",
                    "properties": {
                      "files": { "type": "array", "items": { "type": "string" } },
                      "directories": { "type": "array", "items": { "type": "string" } }
                    },
                    "description": "이 TASK가 건드릴 파일/디렉토리 범위. MAIN/SUB 동시작업 충돌 방지에 사용됨."
                  },
                  "clientRequestId": {
                    "type": "string",
                    "description": "같은 TASK를 실수로 두 번 보내도 중복 생성되지 않도록 하는 idempotency 키. 매 TASK마다 고유값 사용 권장."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "이미 존재하는 clientRequestId — 기존 TASK를 그대로 반환." },
          "201": { "description": "새 TASK 생성됨." },
          "400": { "description": "입력값 오류 (assignee 누락/오타, content 없음, 길이 초과 등)." },
          "401": { "description": "인증 실패." },
          "413": { "description": "요청 본문이 너무 큼." },
          "429": { "description": "요청 빈도 제한 초과." }
        }
      }
    },
    "/api/ai-handoff/messages": {
      "get": {
        "operationId": "get_ai_handoff_status",
        "summary": "TASK/REPORT 타임라인을 시간순으로 조회한다.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 50, "maximum": 200 },
            "description": "최대 반환 개수."
          },
          {
            "name": "after",
            "in": "query",
            "schema": { "type": "string" },
            "description": "이 ISO 시각 이후에 생성된 메시지만 조회 (증분 폴링용)."
          }
        ],
        "responses": {
          "200": { "description": "메시지 목록." },
          "401": { "description": "인증 실패." }
        }
      }
    },
    "/api/ai-handoff/latest-report": {
      "get": {
        "operationId": "get_ai_reports",
        "summary": "가장 최근에 도착한 REPORT 1건을 조회한다.",
        "responses": {
          "200": { "description": "가장 최근 REPORT (없으면 report: null)." },
          "401": { "description": "인증 실패." }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "AiHandoffAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Authorization: Bearer <AI_HANDOFF_API_SECRET>. 이 값은 스톤님이 ChatGPT Action 인증 설정 화면에 직접 입력합니다 — 이 스펙 파일에는 절대 포함하지 않습니다."
      }
    }
  },
  "security": [{ "AiHandoffAuth": [] }]
}